+++ test/templates/template-parts/acf/basic.php
@@ -8,7 +8,7 @@ printf( '<h3>Basic: %s</h3>', $acf_test_post_id );
 <dl>
 	<?php
 
-	foreach ( [ 'text', 'textarea', 'number', 'range', 'e-mail', 'url' ] as $selector ) {
+	foreach ( [ 'text', 'textarea', 'number', 'range', 'e-mail', 'url', 'ambigous_name' ] as $selector ) {
 		?>
 		<dt><?php
 			do_action( 'acf_customizer_field', $selector, $acf_test_post_id );